+++ src/chain/tezos/TezosContractIntrospector.ts
@@ -41,8 +41,8 @@ export namespace TezosContractIntrospector {
      * @returns {Promise<EntryPoint[]>} Information about the entry points, including name, parameters, structure, and invocation parameter generator.
      */
     export async function generateEntryPointsFromAddress(conseilServer: ConseilServerInfo, network: string, contractAddress: string): Promise<EntryPoint[]> {
-        const account: any[] = await TezosConseilClient.getAccount(conseilServer, network, contractAddress);
-        const contractCode: string = account[0].script;
+        const account = await TezosConseilClient.getAccount(conseilServer, network, contractAddress);
+        const contractCode: string = account.script;
         return generateEntryPointsFromCode(contractCode);
     }
 }